Move chat engine out of bancho package and make BanchoUserMap more generic for users
