+++ src/main/kotlin/io/hirasawa/server/HirasawaProject.kt
@@ -1,12 +1,11 @@
 package io.hirasawa.server
 
-import io.hirasawa.server.bancho.chat.ChatChannel
-import io.hirasawa.server.bancho.chat.command.ConsoleCommandSender
+import io.hirasawa.server.chat.ChatChannel
+import io.hirasawa.server.chat.command.ConsoleCommandSender
 import io.hirasawa.server.plugin.internalplugins.InternalBanchoPlugin
 import io.hirasawa.server.plugin.internalplugins.InternalGameApiPlugin
 import io.hirasawa.server.plugin.internalplugins.InternalIrcPlugin
 import io.hirasawa.server.plugin.internalplugins.InternalWebFrontendPlugin
-import io.hirasawa.server.webserver.enums.CommonDomains
 import java.io.File
 
 fun main() {
