--- src/main/kotlin/io/hirasawa/server/bancho/chat/command/ConsoleCommandSender.kt
@@ -1,3 +1,0 @@
-package io.hirasawa.server.bancho.chat.command
-
-class ConsoleCommandSender: CommandSender
