+++ src/main/kotlin/io/hirasawa/server/bancho/packets/ChannelAvailableAutojoinPacket.kt
@@ -1,6 +1,6 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 
 class ChannelAvailableAutojoinPacket(chatChannel: ChatChannel):
         BanchoPacket(BanchoPacketType.BANCHO_CHANNEL_AVAILABLE_AUTOJOIN) {
