+++ src/main/kotlin/io/hirasawa/server/bancho/packets/ChannelAvailablePacket.kt
@@ -1,6 +1,6 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 
 class ChannelAvailablePacket(chatChannel: ChatChannel): BanchoPacket(BanchoPacketType.BANCHO_CHANNEL_AVAILABLE) {
     init {
