+++ src/main/kotlin/io/hirasawa/server/bancho/packets/ChannelJoinSuccessPacket.kt
@@ -1,6 +1,6 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 
 class ChannelJoinSuccessPacket(chatChannel: ChatChannel): BanchoPacket(BanchoPacketType.BANCHO_CHANNEL_JOIN_SUCCESS) {
     init {
