+++ src/main/kotlin/io/hirasawa/server/bancho/packets/ChannelRevokedPacket.kt
@@ -1,6 +1,6 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 
 class ChannelRevokedPacket(channelName: String): BanchoPacket(BanchoPacketType.BANCHO_CHANNEL_REVOKED) {
     constructor(chatChannel: ChatChannel): this(chatChannel.name)
