+++ src/main/kotlin/io/hirasawa/server/bancho/packets/UserPresenceBundlePacket.kt
@@ -1,6 +1,5 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
 import io.hirasawa.server.bancho.serialisation.BanchoIntListWriter
 
 class UserPresenceBundlePacket(list: List<Int>):
