+++ src/main/kotlin/io/hirasawa/server/bancho/packets/UserPresenceSinglePacket.kt
@@ -1,8 +1,5 @@
 package io.hirasawa.server.bancho.packets
 
-import io.hirasawa.server.bancho.chat.ChatChannel
-import io.hirasawa.server.bancho.serialisation.BanchoIntListWriter
-
 class UserPresenceSinglePacket(id: Int):
         BanchoPacket(BanchoPacketType.BANCHO_USER_PRESENCE_SINGLE) {
     init {
