+++ src/main/kotlin/io/hirasawa/server/chat/ChatChannel.kt
@@ -1,7 +1,7 @@
-package io.hirasawa.server.bancho.chat
+package io.hirasawa.server.chat
 
 import io.hirasawa.server.Hirasawa
-import io.hirasawa.server.bancho.chat.message.GlobalChatMessage
+import io.hirasawa.server.chat.message.GlobalChatMessage
 import io.hirasawa.server.bancho.packets.BanchoPacket
 import io.hirasawa.server.bancho.packets.ChannelAvailablePacket
 import io.hirasawa.server.bancho.packets.SendMessagePacket
