+++ src/main/kotlin/io/hirasawa/server/chat/command/ChatCommand.kt
@@ -1,4 +1,4 @@
-package io.hirasawa.server.bancho.chat.command
+package io.hirasawa.server.chat.command
 
 abstract class ChatCommand(val name: String, val description: String, val permission: String = "") {
     abstract fun onCommand(context: CommandContext, command: String, args: List<String>): Boolean
