+++ src/main/kotlin/io/hirasawa/server/chat/command/CommandContext.kt
@@ -1,7 +1,7 @@
-package io.hirasawa.server.bancho.chat.command
+package io.hirasawa.server.chat.command
 
 import io.hirasawa.server.Hirasawa
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 
 data class CommandContext(val sender: CommandSender, val channel: ChatChannel) {
     fun respond(message: String) {
