+++ src/main/kotlin/io/hirasawa/server/chat/command/CommandSender.kt
@@ -1,0 +1,3 @@
+package io.hirasawa.server.chat.command
+
+interface CommandSender
