+++ src/main/kotlin/io/hirasawa/server/chat/command/ConsoleCommandSender.kt
@@ -1,0 +1,3 @@
+package io.hirasawa.server.chat.command
+
+class ConsoleCommandSender: CommandSender
