+++ src/main/kotlin/io/hirasawa/server/chat/message/GlobalChatMessage.kt
@@ -1,6 +1,6 @@
-package io.hirasawa.server.bancho.chat.message
+package io.hirasawa.server.chat.message
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 import io.hirasawa.server.bancho.user.User
 
 class GlobalChatMessage(user: User, val channel: ChatChannel, message: String):
