+++ src/main/kotlin/io/hirasawa/server/chat/message/PrivateChatMessage.kt
@@ -1,6 +1,5 @@
-package io.hirasawa.server.bancho.chat.message
+package io.hirasawa.server.chat.message
 
-import io.hirasawa.server.bancho.chat.message.ChatMessage
 import io.hirasawa.server.bancho.user.User
 
 class PrivateChatMessage(source: User, val destination: User, message: String):
