+++ src/main/kotlin/io/hirasawa/server/commands/UserInfoCommand.kt
@@ -1,12 +1,8 @@
 package io.hirasawa.server.commands
 
 import io.hirasawa.server.Hirasawa
-import io.hirasawa.server.bancho.chat.command.ChatCommand
-import io.hirasawa.server.bancho.chat.command.CommandContext
-import io.hirasawa.server.bancho.chat.command.CommandSender
-import io.hirasawa.server.bancho.chat.message.PrivateChatMessage
-import io.hirasawa.server.bancho.packets.SendMessagePacket
-import io.hirasawa.server.bancho.user.BanchoUser
+import io.hirasawa.server.chat.command.ChatCommand
+import io.hirasawa.server.chat.command.CommandContext
 
 class UserInfoCommand: ChatCommand("userinfo", "Command to give internal information about a user") {
     override fun onCommand(context: CommandContext, command: String, args: List<String>): Boolean {
