+++ src/main/kotlin/io/hirasawa/server/irc/IrcServer.kt
@@ -1,10 +1,6 @@
 package io.hirasawa.server.irc
 
 import io.hirasawa.server.Hirasawa
-import io.hirasawa.server.bancho.chat.message.ChatMessage
-import io.hirasawa.server.bancho.chat.message.GlobalChatMessage
-import io.hirasawa.server.bancho.chat.message.PrivateChatMessage
-import io.hirasawa.server.bancho.user.HirasawaBot
 import io.hirasawa.server.bancho.user.User
 import io.hirasawa.server.irc.clientcommands.*
 import io.hirasawa.server.irc.objects.IrcUser
