+++ src/main/kotlin/io/hirasawa/server/objects/UserMap.kt
@@ -1,4 +1,4 @@
-package io.hirasawa.server.bancho.objects
+package io.hirasawa.server.objects
 
 import io.hirasawa.server.Hirasawa
 import io.hirasawa.server.bancho.user.BanchoUser
@@ -6,42 +6,52 @@ import io.hirasawa.server.bancho.user.User
 import java.util.*
 import kotlin.collections.HashMap
 
-class BanchoUserMap {
+/**
+ * Basic map that contains a list of users that creates an id, username and UUID cache to easily reference them
+ */
+class UserMap<T: User> {
     private val uuidCache = HashMap<UUID, BanchoUser>()
-    private val usernameCache = HashMap<String, BanchoUser>()
-    private val idCache = HashMap<Int, BanchoUser>()
+    private val usernameCache = HashMap<String, T>()
+    private val idCache = HashMap<Int, T>()
 
+    val usernameKeys get() = usernameCache.keys
+    val uuidKeys get() = uuidCache.keys
     val idKeys get() = idCache.keys
+    val values get() = idCache.values
 
     operator fun get(key: UUID): BanchoUser? {
         return uuidCache[key]
     }
 
-    operator fun get(key: String): BanchoUser? {
+    operator fun get(key: String): T? {
         return usernameCache[key]
     }
 
-    operator fun get(key: Int): BanchoUser? {
+    operator fun get(key: Int): T? {
         return idCache[key]
     }
 
-    operator fun get(user: User): BanchoUser? {
+    operator fun get(user: T): T? {
         return idCache[user.id]
     }
 
-    operator fun iterator(): MutableIterator<BanchoUser> {
-        return uuidCache.values.iterator()
+    operator fun iterator(): MutableIterator<T> {
+        return idCache.values.iterator()
     }
 
-    fun add(user: BanchoUser) {
-        uuidCache[user.uuid] = user
+    fun add(user: T) {
+        if (user is BanchoUser) {
+            uuidCache[user.uuid] = user
+        }
         usernameCache[user.username] = user
         idCache[user.id] = user
         Hirasawa.chatEngine.addUser(user)
     }
 
-    fun remove(user: BanchoUser) {
-        uuidCache.remove(user.uuid)
+    fun remove(user: T) {
+        if (user is BanchoUser)  {
+            uuidCache.remove(user.uuid)
+        }
         usernameCache.remove(user.username)
         idCache.remove(user.id)
         Hirasawa.chatEngine.removeUser(user)
@@ -59,11 +69,10 @@ class BanchoUserMap {
         return key in idCache.keys
     }
 
-    operator fun contains(banchoUser: BanchoUser): Boolean {
-        return banchoUser in uuidCache.values
-    }
-
-    operator fun contains(user: User): Boolean {
+    operator fun contains(user: T): Boolean {
+        if (user is BanchoUser) {
+            return user.uuid in uuidCache.keys
+        }
         return user.id in idCache.keys
     }
 }
