+++ src/main/kotlin/io/hirasawa/server/plugin/event/bancho/ChannelJoinEvent.kt
@@ -1,6 +1,6 @@
 package io.hirasawa.server.plugin.event.bancho
 
-import io.hirasawa.server.bancho.chat.ChatChannel
+import io.hirasawa.server.chat.ChatChannel
 import io.hirasawa.server.bancho.user.BanchoUser
 import io.hirasawa.server.plugin.event.Cancelable
 import io.hirasawa.server.plugin.event.HirasawaEvent
