+++ src/main/kotlin/io/hirasawa/server/plugin/event/web/ScoreSubmitEvent.kt
@@ -1,7 +1,5 @@
 package io.hirasawa.server.plugin.event.web
 
-import io.hirasawa.server.bancho.chat.message.ChatMessage
-import io.hirasawa.server.bancho.user.BanchoUser
 import io.hirasawa.server.objects.Score
 import io.hirasawa.server.plugin.event.Cancelable
 import io.hirasawa.server.plugin.event.HirasawaEvent
